Erreur 500 à la création d'une galerie avec un slug vide
