+++ zds/tutorialv2/__init__.py
@@ -1,4 +1,3 @@
 import re
 
 REPLACE_IMAGE_PATTERN = re.compile(r"(?P<start>)(?P<text>!\[.*?\]\()(?P<url>.+?)(?P<end>\))")
-VALID_SLUG = re.compile(r"^[a-z0-9\-_]+$")
